backend/pyproject.toml
[project]
name = "vms-poc-backend"
version = "0.1.0"
description = "Backend service for AI-based VMS (Video Management System) with ONVIF PTZ camera control and streaming features."
authors = [
    { name = "Yash Joshi", email = "yash.joshi@eminds.ai" },
    { name = "Shiv Shankar" }
]

requires-python = ">=3.10"
keywords = ["vms", "onvif", "ptz", "camera", "fastapi", "iot", "streaming"]
license = { text = "MIT" }

dependencies = [
    "onvif-zeep>=0.2.12,<0.3.0",
    "fastapi>=0.115.12,<0.116.0",
    "uvicorn>=0.34.2,<0.35.0",
    "pydantic>=2.11.5,<3.0.0",
    "pydantic-settings>=2.9.1,<3.0.0",
    "opencv-python>=4.11.0.86,<5.0.0.0"
]

[tool.poetry]
packages = [
    { include = "app", from = "src" }
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
